vmupdate: extract reusable utilities to vmupdate/utils.py
Move shutdown_domains, get_feature, get_boolean_feature, and is_stale
out of vmupdate.py into a new vmupdate/utils.py module. These functions
have no dependency on the CLI entry point and are useful from other
modules (e.g. qube_connection.py) without causing circular imports.

Update vmupdate.py and qube_connection.py to import from utils.
